fix: Announcement banner flicker (#2565)
